+++ effect_screening/eligibility.py
@@ -72,7 +72,12 @@ class ScreeningEligibility(BaseScreeningEligibility):
                 "reaction_to_study_drugs",
                 "on_fluconazole",
                 "contraindicated_meds",
-                "mg_ssx_since_crag",
+                "mg_severe_headache",
+                "mg_headache_nuchal_rigidity",
+                "mg_headache_vomiting",
+                "mg_seizures",
+                "mg_gcs_lt_15",
+                "any_other_mg_ssx",
                 "jaundice",
                 "cm_in_csf",
             ]
@@ -89,10 +94,6 @@ class ScreeningEligibility(BaseScreeningEligibility):
             reasons_ineligible.update(cm_in_csf="Positive evidence of CM on CSF")
         if self.model_obj.jaundice not in [NO, NOT_ANSWERED]:
             reasons_ineligible.update(jaundice="Jaundice")
-        if self.model_obj.mg_ssx_since_crag not in [NO, NOT_ANSWERED]:
-            reasons_ineligible.update(
-                mg_ssx_since_crag="Signs of symptomatic meningitis"
-            )
         if self.model_obj.on_fluconazole not in [NO, NOT_ANSWERED]:
             reasons_ineligible.update(on_fluconazole="On fluconazole")
         if self.model_obj.pregnant == YES:
@@ -105,4 +106,19 @@ class ScreeningEligibility(BaseScreeningEligibility):
             reasons_ineligible.update(
                 reaction_to_study_drugs="Serious reaction to flucytosine or fluconazole"
             )
+        reasons_ineligible = self.review_mg_ssx(reasons_ineligible)
+        return reasons_ineligible
+
+    def review_mg_ssx(self, reasons_ineligible: dict) -> dict:
+        """Exclusion for clinical symptoms/signs of symptomatic meningitis."""
+        for mg_ssx in [
+            "mg_severe_headache",
+            "mg_headache_nuchal_rigidity",
+            "mg_headache_vomiting",
+            "mg_seizures",
+            "mg_gcs_lt_15",
+            "any_other_mg_ssx",
+        ]:
+            if getattr(self.model_obj, mg_ssx, None) not in [NO, NOT_ANSWERED]:
+                reasons_ineligible.update({mg_ssx: "Signs of symptomatic meningitis"})
         return reasons_ineligible