+++ effect_screening/tests/tests/test_eligibility.py
@@ -1,3 +1,5 @@
+from typing import Dict
+
 from dateutil.relativedelta import relativedelta
 from django.test import TestCase, tag
 from edc_constants.constants import (
@@ -5,24 +7,25 @@ from edc_constants.constants import (
     MALE,
     NEG,
     NO,
+    NOT_ANSWERED,
     NOT_APPLICABLE,
-    OTHER,
     PENDING,
     POS,
     YES,
 )
 from edc_utils.date import get_utcnow
 
-from effect_lists.models import SiSxMeningitis
 from effect_screening.eligibility import ScreeningEligibility
 from effect_screening.forms import SubjectScreeningForm
 from effect_screening.models import SubjectScreening
-from effect_subject.constants import HEADACHE
 
+from ...choices import POS_NEG_NOT_ANSWERED
 from ..effect_test_case_mixin import EffectTestCaseMixin
 
 
 class TestForms(EffectTestCaseMixin, TestCase):
+    ELIGIBLE_CD4_VALUE = 99
+
     @staticmethod
     def get_data():
         return {
@@ -47,13 +50,14 @@ class TestForms(EffectTestCaseMixin, TestCase):
             "age_in_years": 25,
             "unsuitable_for_study": NO,
             "unsuitable_agreed": NOT_APPLICABLE,
+            "any_other_mg_ssx_other": "",
         }
 
     @property
     def inclusion_criteria(self):
         return dict(
             hiv_pos=YES,
-            cd4_value=99,
+            cd4_value=self.ELIGIBLE_CD4_VALUE,
             cd4_date=(get_utcnow() - relativedelta(days=7)).date(),
             serum_crag_value=POS,
             serum_crag_date=(get_utcnow() - relativedelta(days=6)).date(),
@@ -69,13 +73,25 @@ class TestForms(EffectTestCaseMixin, TestCase):
             contraindicated_meds=NO,
             cm_in_csf=NO,
             cm_in_csf_method=NOT_APPLICABLE,
+            mg_severe_headache=NO,
+            mg_headache_nuchal_rigidity=NO,
+            mg_headache_vomiting=NO,
+            mg_seizures=NO,
+            mg_gcs_lt_15=NO,
+            any_other_mg_ssx=NO,
             jaundice=NO,
-            mg_ssx_since_crag=NO,
             on_fluconazole=NO,
             pregnant=NOT_APPLICABLE,
             breast_feeding=NO,
             prior_cm_epidose=NO,
             reaction_to_study_drugs=NO,
+        )
+
+    def get_valid_opts(self) -> Dict:
+        return dict(
+            **self.inclusion_criteria,
+            **self.exclusion_criteria,
+            **self.get_basic_opts(),
         )
 
     def test_screening_ok(self):
@@ -125,6 +141,47 @@ class TestForms(EffectTestCaseMixin, TestCase):
         obj = ScreeningEligibility(model_obj=model_obj)
         self.assertDictEqual({}, obj.reasons_ineligible)
         self.assertTrue(obj.is_eligible)
+
+    def test_mg_ssx_ineligible(self):
+        for mg_ssx in [
+            "mg_severe_headache",
+            "mg_headache_nuchal_rigidity",
+            "mg_headache_vomiting",
+            "mg_seizures",
+            "mg_gcs_lt_15",
+            "any_other_mg_ssx",
+        ]:
+            with self.subTest(mg_ssx=mg_ssx):
+                model_obj = SubjectScreening.objects.create(
+                    **self.inclusion_criteria,
+                    **self.exclusion_criteria,
+                    **self.get_basic_opts(),
+                )
+                setattr(model_obj, mg_ssx, NOT_ANSWERED)
+                obj = ScreeningEligibility(model_obj=model_obj)
+                self.assertFalse(obj.is_eligible)
+                self.assertDictEqual(
+                    {"exclusion_criteria": "Incomplete exclusion criteria"},
+                    obj.reasons_ineligible,
+                )
+
+                setattr(model_obj, mg_ssx, YES)
+                obj = ScreeningEligibility(model_obj=model_obj)
+                self.assertFalse(obj.is_eligible)
+                self.assertDictEqual(
+                    {mg_ssx: "Signs of symptomatic meningitis"},
+                    obj.reasons_ineligible,
+                )
+
+                setattr(model_obj, mg_ssx, NO)
+                obj = ScreeningEligibility(model_obj=model_obj)
+                self.assertTrue(obj.is_eligible)
+                self.assertDictEqual({}, obj.reasons_ineligible)
+
+    def test_valid_opts_ok(self):
+        form = SubjectScreeningForm(data=self.get_valid_opts())
+        form.is_valid()
+        self.assertDictEqual({}, form._errors)
 
     @tag("preg")
     def test_male_preg_raises(self):
@@ -233,6 +290,53 @@ class TestForms(EffectTestCaseMixin, TestCase):
         self.assertEqual(obj.reasons_ineligible, {})
         self.assertTrue(obj.is_eligible)
 
+    def test_eligible_cd4_values_ok(self):
+        opts = self.get_valid_opts()
+        eligible_cd4_values = [0, 1, 80, 99]
+        for cd4_value in eligible_cd4_values:
+            with self.subTest(cd4_value=cd4_value):
+                opts.update(cd4_value=cd4_value)
+                form = SubjectScreeningForm(data=opts)
+                form.is_valid()
+                self.assertDictEqual({}, form._errors)
+
+    def test_ineligible_cd4_value_raises_validation_error(self):
+        opts = self.get_valid_opts()
+        ineligible_cd4_values = [-1, 100, 120, 200]
+        for cd4_value in ineligible_cd4_values:
+            with self.subTest(cd4_value=cd4_value):
+                opts.update(cd4_value=cd4_value)
+                form = SubjectScreeningForm(data=opts)
+                form.is_valid()
+                self.assertIn("cd4_value", form._errors)
+                self.assertDictEqual(
+                    {
+                        "cd4_value": [
+                            "Ensure this value is less than or equal to 99."
+                            if cd4_value > 0
+                            else "Ensure this value is greater than or equal to 0."
+                        ]
+                    },
+                    form._errors,
+                )
+        opts.update(cd4_value=self.ELIGIBLE_CD4_VALUE)
+        form = SubjectScreeningForm(data=opts)
+        form.is_valid()
+        self.assertDictEqual({}, form._errors)
+
+    def test_cd4_value_required(self):
+        opts = self.get_valid_opts()
+        for cd4_value in [None, ""]:
+            with self.subTest(cd4_value=cd4_value):
+                opts.update(cd4_value=cd4_value)
+
+                form = SubjectScreeningForm(data=opts)
+                form.is_valid()
+                self.assertIn("cd4_value", form._errors)
+                self.assertDictEqual(
+                    {"cd4_value": ["This field is required."]}, form._errors
+                )
+
     def test_cd4_date_within_21_days(self):
         opts = dict(
             **self.inclusion_criteria,
@@ -249,6 +353,64 @@ class TestForms(EffectTestCaseMixin, TestCase):
         form = SubjectScreeningForm(data=opts)
         form.is_valid()
         self.assertNotIn("cd4_date", form._errors)
+        self.assertDictEqual({}, form._errors)
+
+    def test_crag_not_positive_raises_validation_error(self):
+        opts = self.get_valid_opts()
+        for choice in [c[0] for c in POS_NEG_NOT_ANSWERED if c[0] != POS]:
+            with self.subTest(serum_crag_value=choice):
+                opts.update(serum_crag_value=choice)
+                form = SubjectScreeningForm(data=opts)
+                form.is_valid()
+                self.assertIn("serum_crag_value", form._errors)
+                self.assertDictEqual(
+                    {
+                        "serum_crag_value": [
+                            "Invalid. "
+                            "Subject must have positive serum/plasma CrAg test result."
+                        ]
+                    },
+                    form._errors,
+                )
+
+    def test_serum_crag_date_within_14_days(self):
+        opts = self.get_valid_opts()
+        report_datetime = opts.get("report_datetime")
+        cd4_date = report_datetime - relativedelta(days=20)
+
+        opts.update(
+            cd4_date=cd4_date,
+            serum_crag_date=report_datetime - relativedelta(days=14 + 1),
+        )
+        form = SubjectScreeningForm(data=opts)
+        form.is_valid()
+        self.assertIn("serum_crag_date", form._errors)
+        self.assertDictEqual(
+            {
+                "serum_crag_date": [
+                    "Invalid. Cannot be more than 14 days before the report date"
+                ]
+            },
+            form._errors,
+        )
+
+        opts.update(
+            cd4_date=cd4_date,
+            serum_crag_date=report_datetime - relativedelta(days=14),
+        )
+        form = SubjectScreeningForm(data=opts)
+        form.is_valid()
+        self.assertNotIn("cd4_date", form._errors)
+        self.assertDictEqual({}, form._errors)
+
+        opts.update(
+            cd4_date=cd4_date,
+            serum_crag_date=report_datetime - relativedelta(days=14 - 1),
+        )
+        form = SubjectScreeningForm(data=opts)
+        form.is_valid()
+        self.assertNotIn("cd4_date", form._errors)
+        self.assertDictEqual({}, form._errors)
 
     def test_serum_crag_date_not_before_cd4_date(self):
         opts = dict(
@@ -309,34 +471,20 @@ class TestForms(EffectTestCaseMixin, TestCase):
         self.assertNotIn("cm_in_csf", obj.reasons_ineligible)
         self.assertTrue(obj.is_eligible)
 
-    def test_mg_ssx(self):
-        opts = dict(
-            **self.inclusion_criteria,
-            **self.exclusion_criteria,
-            **self.get_basic_opts(),
+    def test_any_other_mg_ssx_other(self):
+        opts = self.get_valid_opts()
+        opts.update(any_other_mg_ssx=YES, any_other_mg_ssx_other="")
+
+        form = SubjectScreeningForm(data=opts)
+        form.is_valid()
+        self.assertIn("any_other_mg_ssx_other", form._errors)
+        self.assertDictEqual(
+            {"any_other_mg_ssx_other": ["This field is required."]}, form._errors
         )
-        opts.update(mg_ssx_since_crag=YES)
+
+        opts.update(any_other_mg_ssx=YES, any_other_mg_ssx_other="Some other sx")
 
         form = SubjectScreeningForm(data=opts)
         form.is_valid()
-        self.assertIn("mg_ssx", form._errors)
-
-        ssx = SiSxMeningitis.objects.filter(name=HEADACHE)
-        opts.update(mg_ssx=ssx)
-
-        form = SubjectScreeningForm(data=opts)
-        form.is_valid()
-        self.assertNotIn("mg_ssx", form._errors)
-
-        ssx = SiSxMeningitis.objects.filter(name=OTHER)
-        opts.update(mg_ssx=ssx)
-
-        form = SubjectScreeningForm(data=opts)
-        form.is_valid()
-        self.assertIn("mg_ssx_other", form._errors)
-
-        opts.update(mg_ssx_other="blah blah")
-
-        form = SubjectScreeningForm(data=opts)
-        form.is_valid()
-        self.assertNotIn("mg_ssx_other", form._errors)
+        self.assertNotIn("any_other_mg_ssx_other", form._errors)
+        self.assertDictEqual({}, form._errors)